fix(lifecycle): tighten NewJobID fallback and surface marshal errors (CodeRabbit #3236484688, #3236484662)
Two related issues on the Phase 5b H3 checkpoint code, both flagged by
CodeRabbit on iqra#56:

1. pkg/lifecycle/checkpoint.go::NewJobID
   The fallback path used '%016x' on time.Now().UnixNano(), but UnixNano
   returns int64. A current nanosecond timestamp formatted as signed hex
   is ~19 chars wide; '%016x' truncates with leading zeros that conflict
   with the high digits, so the fallback could produce ids that fail the
   jobIDPattern regex. Cast to uint64 first so the format emits exactly
   16 lowercase hex chars and never a sign or extra digit. The fallback
   only fires when crypto/rand returns an error, but it must produce a
   valid id when it does, otherwise Save would reject it as 'invalid job
   id' on the way out.

2. batch_checkpoint.go::Run
   The original code discarded the error from json.Marshal(req). If a
   future request type ever carried an unmarshalable field, the discard
   would silently save an empty RequestRaw, and ResumeFromCheckpoint
   would later fail with the misleading 'missing request payload' error.
   Capture the marshal error at job start and propagate it as a typed
   error from Run; this fails the request fast with a useful trace
   rather than corrupting the on-disk checkpoint.

+++ src/services/go-engine/pkg/lifecycle/checkpoint.go
@@ -75,8 +75,11 @@ func NewJobID() string {
 	var b [8]byte
 	if _, err := rand.Read(b[:]); err != nil {
 		// crypto/rand failure is exceptional; fall back to a still-unique-ish
-		// nano-time encoding so the engine never panics on shutdown path.
-		return fmt.Sprintf("%d-%016x", time.Now().Unix(), time.Now().UnixNano())
+		// time-based encoding so the engine never panics on shutdown path.
+		// Use uint64(UnixNano) so %016x produces exactly 16 lowercase hex
+		// chars and never a leading sign or extra digits (a current nano
+		// timestamp is ~19 hex digits when treated as a signed int64).
+		return fmt.Sprintf("%d-%016x", time.Now().Unix(), uint64(time.Now().UnixNano()))
 	}
 	return fmt.Sprintf("%d-%x", time.Now().Unix(), b[:])
 }
